+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.0]
 
-  * Add -v/--verbose option to futhark-doc.
+  * Retain parameter names in output of `futhark-doc`.
+
+  * Add -v/--verbose option to `futhark-doc`.
 
   * Size annotations may now refer to preceding parameters, e.g: